+++ custom.js
@@ -11,6 +11,7 @@ exports.parseNumber        = require('./build/parse').default
 exports.format             = require('./build/format').default
 exports.formatNumber       = require('./build/format').default
 exports.getNumberType      = require('./build/getNumberType').default
+exports.isPossibleNumber   = require('./build/isPossibleNumber').default
 exports.isValidNumber      = require('./build/validate').default
 exports.isValidNumberForRegion = require('./build/isValidNumberForRegion').default