+++ index.d.ts
@@ -43,6 +43,9 @@ export function formatNumber(phone: NationalNumber, country: CountryCode, format
 export function getNumberType(parsedNumber: ParsedNumber): NumberType;
 export function getNumberType(phone: NationalNumber, country?: CountryCode): NumberType;
 
+export function isPossibleNumber(parsedNumber: ParsedNumber): boolean;
+export function isPossibleNumber(phone: NationalNumber, country?: CountryCode): boolean;
+
 export function isValidNumber(parsedNumber: ParsedNumber): boolean;
 export function isValidNumber(phone: NationalNumber, country?: CountryCode): boolean;