+++ test/metadata/1.0.0/isPossibleNumber.test.js
@@ -1,0 +1,16 @@
+import metadata from './metadata.min.json'
+import _isPossibleNumber from '../../../source/isPossibleNumber'
+
+function isPossibleNumber(...parameters)
+{
+	parameters.push(metadata)
+	return _isPossibleNumber.apply(this, parameters)
+}
+
+describe('isPossibleNumber', () =>
+{
+	it('should work in edge cases', function()
+	{
+		expect(() => isPossibleNumber({ phone: '1112223344', country: 'RU' })).to.throw('Metadata too old')
+	})
+})